// unical-support-frontend/frontend/src/app/core/api/models/email-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CategoryResponse } from '../models/category-response';
import { EmailStatus } from '../models/email-status';
export interface EmailResponse {
  body?: (string | null);
  categories?: Array<CategoryResponse>;
  generated_draft?: (string | null);
  gmail_id: string;
  id: string;
  sender: string;
  status: EmailStatus;
  subject?: (string | null);
}